+++ nix/treefmt.nix
@@ -3,7 +3,7 @@
 
   programs.rustfmt = {
     enable = true;
-    edition = "2021";
+    edition = "2024";
     package = pkgs.rust-bin.nightly.latest.default.override {
       extensions = ["rustfmt"];
     };